+++ src/app/service/constants/constants.ts
@@ -1,7 +1,7 @@
 export default Object.freeze({
 	APP_VERSION: process.env.npm_package_version,
 	APP_NAME: 'MadHatter',
-	
+
 	DISCORD_BOT_USER_ID: process.env.DISCORD_BOT_APPLICATION_ID,
 	DB_NAME_DEGEN: 'degen',
 	DB_NAME_NEXTAUTH: 'nextauth',
@@ -19,10 +19,12 @@ export default Object.freeze({
 	DB_COLLECTION_NEXT_AUTH_ACCOUNTS: 'accounts',
 
 	DB_COLLECTION_SCOAP_SQUAD: 'scoapSquad',
+	DB_COLLECTION_SQUAD: 'squad',
+
 	DB_COLLECTION_POAP_ADMINS: 'poapAdmins',
 	DB_COLLECTION_ALLOWLIST: 'allowList',
 	DB_COLLECTION_MEETING_NOTES: 'meetingNotes',
-	
+
 	MONGODB_URI_PARTIAL: `${process.env.MONGODB_PREFIX}://${process.env.MONGODB_USERNAME}:${process.env.MONGODB_PASS}@${process.env.MONGODB_CLUSTER}/`,
 	MONGODB_OPTIONS: '?retryWrites=true&w=majority',